refactor: update `isWeb` imports to be from the `constants` package
