+++ packages/accordion/package.json
@@ -25,6 +25,7 @@
     "@tamagui/collapsible": "1.79.9",
     "@tamagui/collection": "1.79.9",
     "@tamagui/compose-refs": "1.79.9",
+    "@tamagui/constants": "1.79.9",
     "@tamagui/core": "1.79.9",
     "@tamagui/create-context": "1.79.9",
     "@tamagui/polyfill-dev": "1.79.9",
