+++ packages/adapt/package.json
@@ -26,6 +26,7 @@
     }
   },
   "dependencies": {
+    "@tamagui/constants": "1.79.9",
     "@tamagui/core": "1.79.9"
   },
   "devDependencies": {
