+++ packages/alert-dialog/package.json
@@ -32,6 +32,7 @@
     "@tamagui/animate-presence": "1.79.9",
     "@tamagui/aria-hidden": "1.79.9",
     "@tamagui/compose-refs": "1.79.9",
+    "@tamagui/constants": "1.79.9",
     "@tamagui/core": "1.79.9",
     "@tamagui/create-context": "1.79.9",
     "@tamagui/dialog": "1.79.9",
