+++ packages/animations-react-native/package.json
@@ -15,6 +15,7 @@
     "dist"
   ],
   "dependencies": {
+    "@tamagui/constants": "1.79.9",
     "@tamagui/use-presence": "1.79.9",
     "@tamagui/web": "1.79.9"
   },
