+++ packages/font-cherry-bomb/package.json
@@ -22,6 +22,7 @@
     "clean:build": "tamagui-build clean:build"
   },
   "dependencies": {
+    "@tamagui/constants": "1.79.9",
     "@tamagui/core": "1.79.9"
   },
   "devDependencies": {
