+++ packages/font-cherry-bomb/src/index.ts
@@ -1,10 +1,5 @@
-import {
-  FillInFont,
-  GenericFont,
-  createFont,
-  getVariableValue,
-  isWeb,
-} from '@tamagui/core'
+import { isWeb } from '@tamagui/constants'
+import { FillInFont, GenericFont, createFont, getVariableValue } from '@tamagui/core'
 
 export const createCherryBombFont = <A extends GenericFont>(
   font: Partial<A> = {},
