+++ packages/image/package.json
@@ -23,6 +23,7 @@
     "clean:build": "tamagui-build clean:build"
   },
   "dependencies": {
+    "@tamagui/constants": "1.79.9",
     "@tamagui/core": "1.79.9"
   },
   "peerDependencies": {
