+++ packages/popover/package.json
@@ -36,6 +36,7 @@
     "@tamagui/animate": "1.79.9",
     "@tamagui/aria-hidden": "1.79.9",
     "@tamagui/compose-refs": "1.79.9",
+    "@tamagui/constants": "1.79.9",
     "@tamagui/core": "1.79.9",
     "@tamagui/dismissable": "1.79.9",
     "@tamagui/floating": "1.79.9",
