+++ packages/popper/src/Popper.tsx
@@ -1,13 +1,13 @@
 // adapted from radix-ui popper
 
 import { useComposedRefs } from '@tamagui/compose-refs'
+import { isWeb } from '@tamagui/constants'
 import {
   SizeTokens,
   StackProps,
   View as TamaguiView,
   createStyledContext,
   getVariableValue,
-  isWeb,
   styled,
   useIsomorphicLayoutEffect,
   useProps,
