+++ packages/portal/package.json
@@ -27,6 +27,7 @@
     }
   },
   "dependencies": {
+    "@tamagui/constants": "1.79.9",
     "@tamagui/core": "1.79.9",
     "@tamagui/stacks": "1.79.9",
     "@tamagui/use-event": "1.79.9"
