+++ packages/sheet/package.json
@@ -34,6 +34,7 @@
     "@tamagui/animate-presence": "1.79.9",
     "@tamagui/animations-react-native": "1.79.9",
     "@tamagui/compose-refs": "1.79.9",
+    "@tamagui/constants": "1.79.9",
     "@tamagui/core": "1.79.9",
     "@tamagui/create-context": "1.79.9",
     "@tamagui/portal": "1.79.9",
