+++ packages/sheet/src/SheetImplementationCustom.tsx
@@ -1,10 +1,10 @@
 import { AdaptParentContext } from '@tamagui/adapt'
 import { AnimatePresence } from '@tamagui/animate-presence'
 import { useComposedRefs } from '@tamagui/compose-refs'
+import { isWeb } from '@tamagui/constants'
 import {
   Theme,
   getConfig,
-  isWeb,
   themeable,
   useConfiguration,
   useEvent,
