+++ packages/slider/package.json
@@ -32,6 +32,7 @@
   },
   "dependencies": {
     "@tamagui/compose-refs": "1.79.9",
+    "@tamagui/constants": "1.79.9",
     "@tamagui/core": "1.79.9",
     "@tamagui/create-context": "1.79.9",
     "@tamagui/get-token": "1.79.9",
