+++ packages/switch/src/createSwitch.tsx
@@ -1,17 +1,16 @@
 import { useComposedRefs } from '@tamagui/compose-refs'
+import { isWeb } from '@tamagui/constants'
 import {
   NativeValue,
   SizeTokens,
   StackProps,
   TamaguiComponentExpectingVariants,
   composeEventHandlers,
-  isWeb,
   useProps,
   withStaticProperties,
 } from '@tamagui/core'
 import { registerFocusable } from '@tamagui/focusable'
 import { useLabelContext } from '@tamagui/label'
-import { YStack } from '@tamagui/stacks'
 import { useControllableState } from '@tamagui/use-controllable-state'
 import { usePrevious } from '@tamagui/use-previous'
 import * as React from 'react'
@@ -19,6 +18,7 @@ import {
   Switch as NativeSwitch,
   SwitchProps as NativeSwitchProps,
   Platform,
+  View,
 } from 'react-native'
 
 import { SwitchFrame as DefaultSwitchFrame, SwitchThumb } from './Switch'
@@ -241,15 +241,17 @@ export function createSwitch<F extends SwitchComponent, T extends SwitchThumbCom
             }
           })}
         >
-          <YStack
-            alignSelf="stretch"
-            flex={1}
+          <View
+            style={{
+              alignSelf: 'stretch',
+              flex: 1,
+            }}
             onLayout={(e) => {
               setFrameWidth(e.nativeEvent.layout.width)
             }}
           >
             {typeof children === 'function' ? children(checked) : children}
-          </YStack>
+          </View>
         </Frame>
         {isWeb && isFormControl && (
           <BubbleInput
