+++ packages/toggle-group/package.json
@@ -31,6 +31,7 @@
     }
   },
   "dependencies": {
+    "@tamagui/constants": "1.79.9",
     "@tamagui/create-context": "1.79.9",
     "@tamagui/focusable": "1.79.9",
     "@tamagui/font-size": "1.79.9",
